Add experimental PeerGroup interfaces
This is an interface experiment.
I added:
- A new PeerGroup interface
- A new Channel interface
- A new Connection interface

I implemented
- UDP using the new interfaces
- TCP using the new interfaces
- A solution to add addressing as a wrapper

The implementation removes the use of Observables and cache
It also removes the notion of Channel on the server side and now the
user declares handlers to receive messages.

In future commits I intend to add more tests to evaluate how intuitive
the interface feels. I will also attempt to add a notion of removing
already subscribed handlers.