Add onConnectionArrival method to handle incoming connections
In protocols where connections exists (like TCP),
we would like to be able to handle an incoming
connection without waiting for the first message to
arrive through that connection.
In the previous commit, connections were not handled
to the user until the first message arrived. I added
a method called onConnectionArrival to register handlers
for incoming connections. This will allow the user to
close an incoming inactive connection when it arrives.
It will also allow the user to call pre-initialization
code.
Ideally, one would like to have a notion of accepting a
connection. However, netty does not provide this notion.
If we change the underlying library, we could handle the
acceptance process in the new method too.