Provide channels on message reception handlers
In this commit I experiment with the idea of keeping channels in the message reception handlers.
Connections are still kept separate for connection handlers.
The main difference is that connections arrive without messages.
Making this distinction between channels and messages allow to implement both multiplexing and addressing with relatively ease.