Document possible options for addressing
This commit shows three sketch implementations for addressing.
One is based on control messages and two other messages are related to headers.
One of the header approaches separates the notions of Channels and Connections.
The other header approach mixes the notion of a header and the need for a control
message.